+++ resources/views/emails/atrisk.blade.php
@@ -1,0 +1,9 @@
+<html>
+Hello,
+<br><br>
+This is a notification that a report was recently submitted, and it was marked as someone is potentially <strong>at risk of immediate harm</strong>.
+<br><br>
+Please review this report urgently in order to determine if any immediate steps are required to safeguard anyone.
+<br><br>
+You can view the report <a href="{{ config('app.url') }}/report/{{ $id }}">here</a>.
+</html>